+++ packages/parser/__tests__/output.test.ts
@@ -208,6 +208,208 @@ describe('extract to css output pipeline', () => {
     expect(result.css).toMatchInlineSnapshot('""')
   })
 
+  test('string literal - factory', () => {
+    const code = `
+    import { panda } from ".panda/jsx"
+
+    const Example = panda('span')\`
+      color: lightgreen;
+
+      & > strong {
+        color: hotpink;
+      }
+    \`
+
+    const baseStyle = panda.div\`
+        background: transparent;
+        border-radius: 3px;
+        border: 1px solid var(--accent-color);
+        color: token(colors.blue.100);
+        display: inline-block;
+        margin: 0.5rem 1rem;
+        padding: 0.5rem 0;
+        transition: all 200ms ease-in-out;
+        width: 11rem;
+        &:hover {
+          filter: brightness(0.85);
+          &:disabled {
+            filter: brightness(1);
+          }
+        }
+        @media (min-width: 768px) {
+          padding: 1rem 0;
+          &:disabled {
+            filter: brightness(1);
+          }
+        }
+    \`
+     `
+    const result = run(code)
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {
+              " & > strong": {
+                "color": "hotpink",
+              },
+              "color": "lightgreen",
+            },
+          ],
+          "name": "panda",
+          "type": "object",
+        },
+        {
+          "data": [
+            {
+              " &:hover": {
+                " &:disabled": {
+                  "filter": "brightness(1)",
+                },
+                "filter": "brightness(0.85)",
+              },
+              "@media (min-width: 768px)": {
+                " &:disabled": {
+                  "filter": "brightness(1)",
+                },
+                "padding": "1rem 0",
+              },
+              "background": "transparent",
+              "border": "1px solid var(--accent-color)",
+              "border-radius": "3px",
+              "color": "token(colors.blue.100)",
+              "display": "inline-block",
+              "margin": "0.5rem 1rem",
+              "padding": "0.5rem 0",
+              "transition": "all 200ms ease-in-out",
+              "width": "11rem",
+            },
+          ],
+          "name": "panda.div",
+          "type": "object",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer utilities {
+        .text_lightgreen {
+          color: lightgreen
+          }
+
+        .\\\\[\\\\&_\\\\>_strong\\\\]\\\\:text_hotpink > strong {
+          color: hotpink
+              }
+
+        .bg_transparent {
+          background: var(--colors-transparent)
+          }
+
+        .border-radius_3px {
+          border-radius: 3px
+          }
+
+        .border_1px_solid_var\\\\(--accent-color\\\\) {
+          border: 1px solid var(--accent-color)
+          }
+
+        .text_token\\\\(colors\\\\.blue\\\\.100\\\\) {
+          color: var(--colors-blue-100)
+          }
+
+        .d_inline-block {
+          display: inline-block
+          }
+
+        .m_0\\\\.5rem_1rem {
+          margin: 0.5rem 1rem
+          }
+
+        .p_0\\\\.5rem_0 {
+          padding: 0.5rem 0
+          }
+
+        .transition_all_200ms_ease-in-out {
+          transition: all 200ms ease-in-out
+          }
+
+        .w_11rem {
+          width: 11rem
+          }
+
+        .\\\\[\\\\&\\\\:hover\\\\]\\\\:filter_brightness\\\\(0\\\\.85\\\\):hover {
+          filter: brightness(0.85)
+              }
+
+        .\\\\[\\\\&\\\\:hover\\\\]\\\\:\\\\[\\\\&\\\\:disabled\\\\]\\\\:filter_brightness\\\\(1\\\\):hover:disabled {
+          filter: brightness(1)
+                  }
+
+        @media (min-width: 768px) {
+          .\\\\[\\\\@media_\\\\(min-width\\\\:_768px\\\\)\\\\]\\\\:\\\\[\\\\&\\\\:disabled\\\\]\\\\:filter_brightness\\\\(1\\\\):disabled {
+            filter: brightness(1)
+              }
+                  }
+
+        @media (min-width: 768px) {
+          .\\\\[\\\\@media_\\\\(min-width\\\\:_768px\\\\)\\\\]\\\\:p_1rem_0 {
+            padding: 1rem 0
+          }
+              }
+      }"
+    `)
+  })
+
+  test('string literal - css', () => {
+    const code = `
+    import { css } from ".panda/css"
+
+    const className = css\`
+        background: transparent;
+        border-radius: 3px;
+        border: 1px solid var(--accent-color);
+        color: token(colors.blue.100);
+    \`
+     `
+    const result = run(code)
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {
+              "background": "transparent",
+              "border": "1px solid var(--accent-color)",
+              "border-radius": "3px",
+              "color": "token(colors.blue.100)",
+            },
+          ],
+          "name": "css",
+          "type": "object",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer utilities {
+        .bg_transparent {
+          background: var(--colors-transparent)
+          }
+
+        .border-radius_3px {
+          border-radius: 3px
+          }
+
+        .border_1px_solid_var\\\\(--accent-color\\\\) {
+          border: 1px solid var(--accent-color)
+          }
+
+        .text_token\\\\(colors\\\\.blue\\\\.100\\\\) {
+          color: var(--colors-blue-100)
+          }
+      }"
+    `)
+  })
+
   test('runtime conditions', () => {
     const code = `
       import { css } from ".panda/css"
@@ -462,36 +664,117 @@ describe('extract to css output pipeline', () => {
     `)
   })
 
+  test('factory css', () => {
+    const code = `
+    import { panda } from ".panda/jsx"
+
+    // PropertyAccess factory css
+    panda.div({
+      color: "red.100",
+    })
+
+    // CallExpression factory css
+    panda("div", {
+        color: "yellow.100",
+    })
+
+    // TaggedTemplateExpression factory css
+    panda.div\`
+      color: var(--colors-purple-100);
+    \`
+   `
+    const result = run(code)
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {
+              "color": "red.100",
+            },
+          ],
+          "name": "panda.div",
+          "type": "object",
+        },
+        {
+          "data": [
+            {
+              "color": "var(--colors-purple-100)",
+            },
+          ],
+          "name": "panda.div",
+          "type": "object",
+        },
+        {
+          "data": [
+            {
+              "color": "yellow.100",
+            },
+          ],
+          "name": "panda",
+          "type": "object",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer utilities {
+        .text_red\\\\.100 {
+          color: var(--colors-red-100)
+          }
+
+        .text_var\\\\(--colors-purple-100\\\\) {
+          color: var(--colors-purple-100)
+          }
+
+        .text_yellow\\\\.100 {
+          color: var(--colors-yellow-100)
+          }
+      }"
+    `)
+  })
+
   test('cva and factory recipes', () => {
     const code = `
       import { panda } from ".panda/jsx"
       import { cva } from ".panda/css"
 
-      const buttonRecipe = cva({
-        base: {
-          color: "red.100",
-          bg: "red.900",
-        }
-      })
-
-      const Button = panda('button', {
-        base: {
-          color: "green.100",
-          bg: "green.900",
-        }
-      })
-
-      const Input = panda.input({
+      // PropertyAccess factory inline recipe
+      panda.div({
         base: {
           color: "blue.100",
-          bg: "blue.900",
+        },
+        variants: {
+          //
+        }
+      })
+
+      // CallExpression factory inline recipe
+      panda("div", {
+        base: {
+          color: "green.100",
+        },
+        variants: {
+          //
+        }
+      })
+
+      // PropertyAccess factory + cva
+      panda.div(cva({
+        base: {
+          color: "rose.100",
+        },
+      }))
+
+      const buttonRecipe = cva({
+        base: {
+          color: "sky.100",
+          bg: "red.900",
         }
       })
 
       function App () {
         return (
           <>
-            <Button>Click me</Button>
             <Input />
           </>
         )
@@ -502,10 +785,40 @@ describe('extract to css output pipeline', () => {
       [
         {
           "data": [
+            {},
+          ],
+          "name": "panda.div",
+          "type": "object",
+        },
+        {
+          "data": [
             {
               "base": {
-                "bg": "red.900",
-                "color": "red.100",
+                "color": "blue.100",
+              },
+              "variants": {},
+            },
+          ],
+          "name": "panda.div",
+          "type": "cva",
+        },
+        {
+          "data": [
+            {
+              "base": {
+                "color": "green.100",
+              },
+              "variants": {},
+            },
+          ],
+          "name": "panda",
+          "type": "cva",
+        },
+        {
+          "data": [
+            {
+              "base": {
+                "color": "rose.100",
               },
             },
           ],
@@ -516,32 +829,13 @@ describe('extract to css output pipeline', () => {
           "data": [
             {
               "base": {
-                "bg": "green.900",
-                "color": "green.100",
+                "bg": "red.900",
+                "color": "sky.100",
               },
             },
           ],
-          "name": "panda",
-          "type": "cva",
-        },
-        {
-          "data": [
-            {
-              "base": {
-                "bg": "blue.900",
-                "color": "blue.100",
-              },
-            },
-          ],
-          "name": "panda.input",
-          "type": "cva",
-        },
-        {
-          "data": [
-            {},
-          ],
-          "name": "Button",
-          "type": "jsx",
+          "name": "cva",
+          "type": "object",
         },
         {
           "data": [
@@ -555,28 +849,24 @@ describe('extract to css output pipeline', () => {
 
     expect(result.css).toMatchInlineSnapshot(`
       "@layer utilities {
-        .text_red\\\\.100 {
-          color: var(--colors-red-100)
-          }
-
-        .bg_red\\\\.900 {
-          background: var(--colors-red-900)
+        .text_blue\\\\.100 {
+          color: var(--colors-blue-100)
           }
 
         .text_green\\\\.100 {
           color: var(--colors-green-100)
           }
 
-        .bg_green\\\\.900 {
-          background: var(--colors-green-900)
+        .text_rose\\\\.100 {
+          color: var(--colors-rose-100)
           }
 
-        .text_blue\\\\.100 {
-          color: var(--colors-blue-100)
+        .text_sky\\\\.100 {
+          color: var(--colors-sky-100)
           }
 
-        .bg_blue\\\\.900 {
-          background: var(--colors-blue-900)
+        .bg_red\\\\.900 {
+          background: var(--colors-red-900)
           }
       }"
     `)