+++ extension/shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pandacss/extension-shared
 
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies [60df9bd1]
+- Updated dependencies [ead9eaa3]
+  - @pandacss/shared@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes