+++ packages/config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pandacss/config
 
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies [ead9eaa3]
+- Updated dependencies [3a87cff8]
+  - @pandacss/types@0.5.0
+  - @pandacss/preset-panda@0.5.0
+  - @pandacss/preset-base@0.5.0
+  - @pandacss/error@0.5.0
+  - @pandacss/logger@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes