+++ packages/extractor/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @pandacss/extractor
 
+## 0.5.0
+
+### Minor Changes
+
+- ead9eaa3: Add support for tagged template literal version.
+
+  This features is pure css approach to writing styles, and can be a great way to migrate from styled-components and
+  emotion.
+
+  Set the `syntax` option to `template-literal` in the panda config to enable this feature.
+
+  ```js
+  // panda.config.ts
+  export default defineConfig({
+    //...
+    syntax: 'template-literal',
+  })
+  ```
+
+  > For existing projects, you might need to run the `panda codegen --clean`
+
+  You can also use the `--syntax` option to specify the syntax type when using the CLI.
+
+  ```sh
+  panda init -p --syntax template-literal
+  ```
+
+  To get autocomplete for token variables, consider using the
+  [CSS Var Autocomplete](https://marketplace.visualstudio.com/items?itemName=phoenisx.cssvar) extension.
+
+### Patch Changes
+
+- @pandacss/logger@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes